Oceananigans.jl Documentation

Oceananigans.jl Documentation

Grids

Oceananigans.RegularCartesianGrid — Type.

RegularCartesianGrid{T::AbstractFloat} <: Grid

A Cartesian grid with regularly spaces cells and faces so that $Δx$, $Δy$, and $Δz$ are constants. Fields are stored using floating-point values of type T.

source

Oceananigans.RegularCartesianGrid — Method.

RegularCartesianGrid(N, L; dim=3, FloatType=Float64)

Create a regular Cartesian grid with size $N = (N_x, N_y, N_z)$ and domain size $L = (L_x, L_y, L_z)$ where fields are stored using floating-point values of type T.

Examples

julia> g = RegularCartesianGrid((16, 16, 8), (2π, 2π, 2π))

source

Fields

Oceananigans.CellField — Type.

CellField{T<:AbstractArray} <: Field

A cell-centered field defined on a grid G whose values are stored as floating-point values of type T.

source

Oceananigans.FaceFieldX — Type.

FaceFieldX{T<:AbstractArray} <: FaceField

An x-face-centered field defined on a grid G whose values are stored as floating-point values of type T.

source

Oceananigans.FaceFieldY — Type.

FaceFieldY{T<:AbstractArray} <: FaceField

A y-face-centered field defined on a grid G whose values are stored as floating-point values of type T.

source

Oceananigans.FaceFieldZ — Type.

FaceFieldZ{T<:AbstractArray} <: FaceField

A z-face-centered field defined on a grid G whose values are stored as floating-point values of type T.

source

Oceananigans.CellField — Method.

CellField(grid::Grid)

Construct a CellField whose values are defined at the center of a cell.

source

Oceananigans.FaceFieldX — Method.

FaceFieldX(grid::Grid)

A Field whose values are defined on the x-face of a cell.

source

Oceananigans.FaceFieldY — Method.

FaceFieldY(grid::Grid)

A Field whose values are defined on the y-face of a cell.

source

Oceananigans.FaceFieldZ — Method.

FaceFieldZ(grid::Grid)

A Field whose values are defined on the z-face of a cell.

source

Operators

Difference operators

Oceananigans.Operators.δx! — Method.

δx!(g::RegularCartesianGrid, f::CellField, δxf::FaceField)

Compute the difference $\delta_x(f) = f_E - f_W$ between the eastern and western cells of a cell-centered field f and store it in a face-centered field δxf, assuming both fields are defined on a regular Cartesian grid g with periodic boundary condition in the $x$-direction.

source

Oceananigans.Operators.δx! — Method.

δx!(g::RegularCartesianGrid, f::FaceField, δxf::CellField)

Compute the difference $\delta_x(f) = f_E - f_W$ between the eastern and western faces of a face-centered field f and store it in a cell-centered field δxf, assuming both fields are defined on a regular Cartesian grid g with periodic boundary conditions in the $x$-direction.

source

Oceananigans.Operators.δy! — Method.

δy!(g::RegularCartesianGrid, f::CellField, δyf::FaceField)

Compute the difference $\delta_y(f) = f_N - f_S$ between the northern and southern cells of a cell-centered field f and store it in a face-centered field δyf, assuming both fields are defined on a regular Cartesian grid g with periodic boundary condition in the $y$-direction.

source

Oceananigans.Operators.δy! — Method.

δy!(g::RegularCartesianGrid, f::FaceField, δyf::CellField)

Compute the difference $\delta_y(f) = f_N - f_S$ between the northern and southern faces of a face-centered field f and store it in a cell-centered field δyf, assuming both fields are defined on a regular Cartesian grid g with periodic boundary condition in the $y$-direction.

source

Oceananigans.Operators.δz! — Method.

δz!(g::RegularCartesianGrid, f::CellField, δzf::FaceField)

Compute the difference $\delta_z(f) = f_T - f_B$ between the top and bottom cells of a cell-centered field f and store it in a face-centered field δzf, assuming both fields are defined on a regular Cartesian grid g with Neumann boundary condition in the $z$-direction.

source

Oceananigans.Operators.δz! — Method.

δz!(g::RegularCartesianGrid, f::FaceField, δzf::CellField)

Compute the difference $\delta_z(f) = f_T - f_B$ between the top and bottom faces of a face-centered field f and store it in a cell-centered field δzf, assuming both fields are defined on a regular Cartesian grid g with Neumann boundary condition in the $z$-direction.

source

Averaging operators

Oceananigans.Operators.avgx! — Method.

avgx(g::RegularCartesianGrid, f::CellField, favgx::FaceField)

Compute the average $\overline{\;f\;}^x = \frac{f_E + f_W}{2}$ between the eastern and western cells of a cell-centered field f and store it in a g face-centered field favgx, assuming both fields are defined on a regular Cartesian grid g with periodic boundary conditions in the $x$-direction.

source

Divergence operators

Building on top of the differencing operators we can define operators that compute the divergence

\[\nabla\cdotp\mathbf{f} = \frac{1}{V} \left[ \delta_x \left( A_x f_x \right) + \delta_y\left( A_y f_y \right) + \delta_z\left( A_z f_z \right)\right]\]

<!– @docs div!(g::RegularCartesianGrid, fx::FaceFieldX, fy::FaceFieldY, fz::FaceFieldZ, δfx::CellField, δfy::CellField, δfz::CellField, div::CellField) –>