Home

Oceananigans.jl Documentation

Grids

Oceananigans.RegularCartesianGrid — Type.

RegularCartesianGrid

A Cartesian grid with regularly spaces cells and faces so that $Δx$, $Δy$, and $Δz$ are constants. Fields are stored using floating-point values of type T.

source

Oceananigans.RegularCartesianGrid — Method.

RegularCartesianGrid(metadata::ModelMetadata, N, L)

Create a regular Cartesian grid with size $N = (N_x, N_y, N_z)$ and domain size $L = (L_x, L_y, L_z)$ where fields are stored using floating-point values of type T.

Examples

julia> g = RegularCartesianGrid((16, 16, 8), (2π, 2π, 2π))

source

Fields

Oceananigans.CellField — Type.

CellField{T,G<:Grid{T}} <: Field{G}

A cell-centered field defined on a grid G whose values are stored as floating-point values of type T.

source

Oceananigans.FaceFieldX — Type.

FaceFieldX{T,G<:Grid{T}} <: FaceField{G}

An x-face-centered field defined on a grid G whose values are stored as floating-point values of type T.

source

Oceananigans.FaceFieldY — Type.

FaceFieldY{T,G<:Grid{T}} <: FaceField{G}

A y-face-centered field defined on a grid G whose values are stored as floating-point values of type T.

source

Oceananigans.FaceFieldZ — Type.

FaceFieldZ{T,G<:Grid{T}} <: FaceField{G}

A z-face-centered field defined on a grid G whose values are stored as floating-point values of type T.

source

Operators

Difference operators

δx!(g::RegularCartesianGrid, f::CellField, δxf::FaceField)
δx!(g::RegularCartesianGrid, f::FaceField, δxf::CellField)
δy!(g::RegularCartesianGrid, f::CellField, δyf::FaceField)
δy!(g::RegularCartesianGrid, f::FaceField, δyf::CellField)
δz!(g::RegularCartesianGrid, f::CellField, δzf::FaceField)
δz!(g::RegularCartesianGrid, f::FaceField, δzf::CellField)

Averaging operators

avgx!(g::RegularCartesianGrid, f::CellField, favgx::FaceField)

Divergence operators

Building on top of the differencing operators we can define operators that compute the divergence

\[\nabla\cdotp\mathbf{f} = \frac{1}{V} \left[ \delta_x \left( A_x f_x \right) + \delta_y\left( A_y f_y \right) + \delta_z\left( A_z f_z \right)\right]\]

<!– @docs div!(g::RegularCartesianGrid, fx::FaceFieldX, fy::FaceFieldY, fz::FaceFieldZ, δfx::CellField, δfy::CellField, δfz::CellField, div::CellField) –>